backend/requirements_chatbot.txt
groq
langchain-groq
langchain_community
python-dotenv
pymongo
qdrant-client
langchain_google_genai
fitz